sql/schemachanger: move scsqldeps code into scdeps
I couldn't figure out why this package exists.

Release note: None
