+++ pkg/ccl/partitionccl/BUILD.bazel
@@ -17,7 +17,7 @@ go_library(
         "//pkg/sql/pgwire/pgcode",
         "//pkg/sql/pgwire/pgerror",
         "//pkg/sql/rowenc",
-        "//pkg/sql/schemachanger/scsqldeps",
+        "//pkg/sql/schemachanger/scdeps",
         "//pkg/sql/sem/tree",
         "//pkg/sql/types",
         "//pkg/util/encoding",
