+++ pkg/sql/schemachanger/scdeps/BUILD.bazel
@@ -5,7 +5,9 @@ go_library(
     srcs = [
         "backfill_tracker.go",
         "build_deps.go",
+        "ccl_deps.go",
         "exec_deps.go",
+        "index_validator.go",
         "periodic_progress_flusher.go",
         "run_deps.go",
     ],
@@ -18,6 +20,7 @@ go_library(
         "//pkg/kv",
         "//pkg/roachpb:with-mocks",
         "//pkg/security",
+        "//pkg/settings",
         "//pkg/settings/cluster",
         "//pkg/sql/backfill",
         "//pkg/sql/catalog",
@@ -25,12 +28,16 @@ go_library(
         "//pkg/sql/catalog/descpb",
         "//pkg/sql/catalog/descs",
         "//pkg/sql/catalog/resolver",
+        "//pkg/sql/catalog/tabledesc",
+        "//pkg/sql/parser",
         "//pkg/sql/schemachanger/scbuild",
         "//pkg/sql/schemachanger/scexec",
         "//pkg/sql/schemachanger/scexec/scmutationexec",
+        "//pkg/sql/schemachanger/scpb",
         "//pkg/sql/schemachanger/scrun",
         "//pkg/sql/sem/tree",
         "//pkg/sql/sessiondata",
+        "//pkg/sql/sqlerrors",
         "//pkg/sql/sqlutil",
         "//pkg/sql/types",
         "//pkg/util",
