--- pkg/sql/schemachanger/scsqldeps/BUILD.bazel
@@ -1,29 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "scsqldeps",
-    srcs = [
-        "ccl_deps.go",
-        "index_validator.go",
-    ],
-    importpath = "github.com/cockroachdb/cockroach/pkg/sql/schemachanger/scsqldeps",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//pkg/keys",
-        "//pkg/kv",
-        "//pkg/settings",
-        "//pkg/settings/cluster",
-        "//pkg/sql/catalog",
-        "//pkg/sql/catalog/descpb",
-        "//pkg/sql/catalog/tabledesc",
-        "//pkg/sql/parser",
-        "//pkg/sql/schemachanger/scexec",
-        "//pkg/sql/schemachanger/scexec/scmutationexec",
-        "//pkg/sql/schemachanger/scpb",
-        "//pkg/sql/sem/tree",
-        "//pkg/sql/sessiondata",
-        "//pkg/sql/sqlerrors",
-        "//pkg/sql/sqlutil",
-        "@com_github_cockroachdb_errors//:errors",
-    ],
-)
